refactor: rename interpreter sandbox to stdlib
