--- src/core/interpreter/sandbox/index.ts
@@ -1,1 +1,0 @@
-export * from './stdlib';
